+++ ChangeLog.md
@@ -3,3 +3,8 @@
 ## 1.0
 
 Initial version
+
+## 1.1
+
+Bon de préparation pour plusieurs expéditions (picking)
+(Accessible comme action sur plusieurs expéditions, option activable dans le setup du module)